requirements.txt
Package                           Version
--------------------------------- ---------
asgiref                           3.4.1
backports.entry-points-selectable 1.1.0
certifi                           2021.10.8
charset-normalizer                2.0.7
distlib                           0.3.3
Django                            3.2.8
djangorestframework               3.12.4
filelock                          3.3.1
idna                              3.3
pip                               21.3
platformdirs                      2.4.0
pytz                              2021.3
requests                          2.26.0
setuptools                        57.4.0
six                               1.16.0
sqlparse                          0.4.2
urllib3                           1.26.7
virtualenv                        20.8.1
